Move intro to Locations
