+++ js/modules/HomeScene/World/World.js
@@ -19,7 +19,6 @@ export default class World extends EventEmitter {
         this.camera = this.homeScene.camera;
         this.resources = this.homeScene.resources;
 
-        this.intro = this.intro.bind(this);
         this.setup = this.setup.bind(this);
 
         this.resources.on("ready", this.setup);
@@ -29,14 +28,13 @@ export default class World extends EventEmitter {
         if (this.parallaxGroup) {
             this.parallaxGroup.update();
         }
-
         if (this.locations) {
             this.locations.update();
         }
     }
 
     resize() {
-        if (this.locations?.model) {
+        if (this.locations) {
             this.locations.resize();
         }
     }
@@ -45,74 +43,5 @@ export default class World extends EventEmitter {
         this.parallaxGroup = new ParallaxGroup();
         this.locations = new Locations();
         this.environment = new Environment();
-
-        if (this.debug.active) {
-            this.debug.ui
-                .add({ intro: this.intro }, "intro")
-                .name("Play Intro");
-        } else {
-            this.intro();
-        }
-    }
-
-    intro() {
-        const {
-            camera: { instance: camera },
-            locations: { model },
-        } = this;
-
-        const toScale = model.scale.clone();
-
-        const tl = gsap.timeline({
-            paused: true,
-            onComplete: () => {
-                toScale.dispose();
-            },
-        });
-        tl.fromTo(
-            model.scale,
-            {
-                x: 0,
-                y: 0,
-                z: 0,
-            },
-            {
-                x: toScale.x,
-                y: toScale.y,
-                z: toScale.z,
-                duration: 1,
-                ease: "power2.inOut",
-            },
-            0
-        );
-        tl.fromTo(
-            camera.position,
-            {
-                x: 2,
-                y: 0,
-                z: 2,
-            },
-            {
-                x: 0,
-                y: 0,
-                z: 2,
-                duration: 2,
-                ease: "power2.inOut",
-            },
-            0
-        );
-        tl.fromTo(
-            model.rotation,
-            { x: 0, y: 0, z: 0 },
-            {
-                x: Math.PI * 0.5,
-                y: 0,
-                z: 0,
-                duration: 2,
-                ease: "power2.inOut",
-            },
-            0
-        );
-        tl.play();
     }
 }
